[ARM] msm: change the hw3d driver to let appications directly map the gpu
If another user has control of the device, open() will attempt to revoke the
gpu by releasing a blocked "wait for revoke" ioctl, if any.
If a grace period expires and the gpu is still not relased, the application
is SIGKILLed. On suspend, the same approach is applied to make the
application release the gpu so that hw can be put into low power mode.

Change-Id: Iceae571d3c5f13c04778c104ad54ccd48cb55d7d
Signed-off-by: Dima Zavin <dima@android.com>