[ARM] mach-msm: nand: Add Virtual kpanic mtd partition support.
Adds an option to snag a portion of a specified
partition into a virtual partition named 'kpanic' Useful
for platforms which do not have a kpanic partition but want
to take advantage of CONFIG_APANIC logging. As this is not a real
nand partition, both the snagged and new partitions should be
erased when switching to/from a kernel with this option enabled/disabled.

**** WARNING ****
DO NOT enable if you are using a Recovery/OTA scheme in which
the recovery kernel needs to access the partition you are snagging
from. If you do not heed this warning, you may end up with a
corrupted OTA image and there be dragons. Just don't do it.
**** WARNING ****

Signed-off-by: San Mehat <san@google.com>