[ARM] msm: Add the yamato GPU driver found on qsd8x50
This driver authored and provided by Qualcomm Inc.

Signed-off-by: Dima Zavin <dima@android.com>

msm_kgsl: report rev 2.1.1 instead of 2.1 due to hardware bug

Hardware rev 2.1.1 is incorrectly reported as 2.1 when reading
the chip ID register. This causes triangle strips to be made
larger by the usermode driver and thus reduces triangle rate
in applications such as Powerlift. The chip ID is changed back
to 2.1.1 to avoid this.

CRs-fixed: 187187
Signed-off-by: Simon Wilson <swilson@quicinc.com>

msm_kgsl: Clean up memory when timestamps are freed

Call yamato_runpending when a timestamp is freed to ensure that
the memory actually gets cleaned.

Signed-off-by: Jordan Crouse <jcrouse@quicinc.com>

[ARM] video: msm: kgsl: Fix typo when tearing down contexts on file release

Signed-off-by: Dima Zavin <dima@android.com>

[ARM] video: msm: kgsl: Disable/enable clocks and irq at suspend/resume.

Signed-off-by: Dima Zavin <dima@android.com>

[ARM] msm: kgsl: Integrate latest driver from Qualcomm.

Fixes several context and command stream management issues.
Fix MMU support, and other fixups.

Change-Id: I384694bd611213e7a14a729f55a35d4e7f921d15
Signed-off-by: Dima Zavin <dima@android.com>

[ARM] msm: kgsl: check the right timestamp on drain

Change-Id: I5932bf639ff33c31bdb5686dd63929df3a6b2868
Signed-off-by: Dima Zavin <dima@android.com>

kgsl: request 128MHz EBI1 clock when active

Signed-off-by: Brian Swetland <swetland@google.com>

[ARM] video: msm: kgsl: Add an idle status check function

Add a function that checks if the ringbuffer is empty and all the blocks
in the GPU are idle.

Change-Id: I9d58a2f607f6f7f28d0fbef6d7104e2444f3ebab
Signed-off-by: Dima Zavin <dima@android.com>

[ARM] video: msm: kgsl: disable the GPU during periods of inactivity

The GPU is only enabled when there are pending commands in the ringbuffer.
Whenever the last user of the driver leaves the kernel, a timer is started
which will check the emptyness of the ringbuffer and whether or not the GPU
itself is idle. If the conditions are not met, the timer is restarted. We try
aggressively turn off the GPU after slightly more than a frame of inactivity.

Change-Id: I9a9be81d52431ef2defdc22a3df3af2b2b1ce817
Signed-off-by: Dima Zavin <dima@android.com>

[ARM] video: msm: kgsl: Dump register state on timeout for waittimestamp.

Also, if the user asks for no timeout, force it to be 10 seconds so we
don't end up blocked forever if GPU hangs.

Change-Id: Ieb3ed7c87904b0897997b0b55ba758030d388aed
Signed-off-by: Dima Zavin <dima@android.com>

[ARM] video: msm: kgsl: Write the error register on axi errors

Change-Id: I0536874ddc5f9844b0fe5b6f32457a4470075b1d
Signed-off-by: Dima Zavin <dima@android.com>

[ARM] video: msm: kgsl: GPU is truly idle when rbbm_status is 0x110

Change-Id: I1066d8bab2b9fb8f06a89f120a7467241d64b78b
Signed-off-by: Dima Zavin <dima@android.com>